[NFC] compositing: move InitialCompositorState back to compositing
