+++ components/compositing_traits/lib.rs
@@ -215,24 +215,3 @@ impl Debug for Msg {
         }
     }
 }
-
-/// Data used to construct a compositor.
-pub struct InitialCompositorState {
-    /// A channel to the compositor.
-    pub sender: CompositorProxy,
-    /// A port on which messages inbound to the compositor can be received.
-    pub receiver: CompositorReceiver,
-    /// A channel to the constellation.
-    pub constellation_chan: Sender<ConstellationMsg>,
-    /// A channel to the time profiler thread.
-    pub time_profiler_chan: time::ProfilerChan,
-    /// A channel to the memory profiler thread.
-    pub mem_profiler_chan: mem::ProfilerChan,
-    /// Instance of webrender API
-    pub webrender: webrender::Renderer,
-    pub webrender_document: DocumentId,
-    pub webrender_api: RenderApi,
-    pub webrender_surfman: WebrenderSurfman,
-    pub webrender_gl: Rc<dyn gleam::gl::Gl>,
-    pub webxr_main_thread: webxr::MainThreadRegistry,
-}
